refactored repeated functions, cleaned root folder, put all styles in one file, fixed sortPlanes to be instant
